+++ automod/cachestore/cachestore_redis.go
@@ -1,4 +1,4 @@
-package automod
+package cachestore
 
 import (
 	"context"
